+++ src/components/editor/PostEditor.tsx
@@ -11,6 +11,7 @@ import "@toast-ui/editor/dist/i18n/ko-kr";
 import CategoryDropdown from "@/components/Dropdown/CategoryDropdown";
 import type { MainCategory } from "@/constants/categories";
 import { CATEGORY_MAP } from "@/constants/categories";
+import { useEditorImageUpload } from "@/hooks/useEditorImageUpload";
 
 export default function PostEditor() {
   const editorRef = useRef<Editor>(null);
@@ -102,49 +103,8 @@ export default function PostEditor() {
     }
   }, [subCategory, mainCategory]);
 
-  // 이미지 업로드 구현 (S3 Presigned URL)
-  useEffect(() => {
-    const editor = editorRef.current?.getInstance();
-    if (!editor) return;
-  
-    // 중복 방지
-    editor.removeHook("addImageBlobHook");
-  
-    editor.addHook("addImageBlobHook", async (blob: Blob, callback: (url: string, alt: string) => void) => {
-      try {
-        const ext = blob.type.split("/")[1]; // "image/png" → "png"
-        const type = "post"; // post 또는 profile
-  
-        // presigned URL 요청
-        const res = await fetch(
-          `/api/presigned-url?type=${type}&ext=${ext}`,
-          { method: "GET" }
-        );
-  
-        if (!res.ok) throw new Error("프리사인드 URL 요청 실패");
-  
-        const json = await res.json();
-        const { url, key } = json.data;
-  
-        // S3에 이미지 PUT 업로드
-        await fetch(url, {
-          method: "PUT",
-          body: blob,
-          headers: {
-            "Content-Type": blob.type,
-          },
-        });
-  
-        // S3 public 이미지 URL을 editor에 삽입
-        const imageUrl = `https://d1eni2d3ighqku.cloudfront.net/${key}`;
-        callback(imageUrl, "이미지");
-  
-      } catch (err) {
-        console.error("이미지 업로드 오류", err);
-        alert("이미지 업로드 중 문제가 발생했습니다.");
-      }
-    });
-  }, [editorRef]);
+  // 이미지 업로드 관련 Hook
+  useEditorImageUpload(editorRef);
 
   // 자동 임시 저장
   useEffect(() => {
